Version 1.3.5.0
* Removed: VLD for compatibility issues with VS2017
* Fixed: Synchronicity issues caused by memory allocation method
* Added: Indicator in status bar for connection status
* Fixed: Bug in output path directory
* Fixed: Connection attempt interval
* Added: Beam Calibration and reference circle
* Added: Button for Beam Calibration / Overlay
* Changed: Appearance of Eraser Button
* Changed: Return pointer in template operators to increase performance
* Added: Template operator * for coords
* Changed: Default values for Box Size
* Fixed: Pupil Box Size < 30 caused Runtime Error.
* Changed: Increased precision of mapping factor
* Improved: tidied up some code

+++ PupilTracker/OffsetTracker.h
@@ -25,7 +25,7 @@ public:
 private:
 
 	double		magnif;		// magnification factor
-	float		conv;		// mapping factor
+	double		conv;		// mapping factor
 
 	COLORREF	bkgrndCol;
 	HPEN		hPenGreen, hPenGrey, hPenWht, hPenClay;
